# PERF: Series.to_frame