--- a/pandas/core/internals/array_manager.py
+++ b/pandas/core/internals/array_manager.py
@@ -1294,6 +1294,15 @@ class SingleArrayManager(BaseArrayManager, SingleDataManager):
         """
         self.arrays[0] = values
 
+    def to_2d_mgr(self, columns: Index) -> ArrayManager:
+        """
+        Manager analogue of Series.to_frame
+        """
+        arrays = [self.arrays[0]]
+        axes = [self.axes[0], columns]
+
+        return ArrayManager(arrays, axes, verify_integrity=False)
+
 
 class NullArrayProxy:
     """