```
import pandas as pd

ser = pd.Series(range(5))

%timeit ser.to_frame()

%prun -s cumtime for n in range(10000): ser.to_frame()
9.71 µs ± 501 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)   # <- PR
33.2 µs ± 651 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- master

```